requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow>=10.0.0
# PostgreSQL için (opsiyonel):
# psycopg2-binary==2.9.9
# python-decouple==3.8
